chore: fix no_std builds
I’m not entirely sure if it’s no_std specifically that causes the
issue, but if ibc-apps-transfer is build without std feature but with
serde feature it enables serde_json/std feature while serde_json
dependency isn’t enabled.  This seems to confuse Cargo which thinks
serde_json is supposed to be a feature.  This all leads to a failure
in dependency resolution:

    error: failed to select a version for `ibc-app-transfer`.
        ... required by package `ibc-apps v0.48.0`
        ... which satisfies dependency `ibc-apps = "^0.48.0"` (locked to 0.48.0) of package `ibc v0.48.0`
        ... which satisfies dependency `ibc = "^0.48.0"` (locked to 0.48.0) of package `ibc-testkit v0.48.0`
        ... which satisfies dependency `ibc-testkit = "^0.48.0"` (locked to 0.48.0) of package …
    versions that meet the requirements `^0.48.0` (locked to 0.48.0) are: 0.48.0

    the package `ibc-apps` depends on `ibc-app-transfer`, with features: `serde_json` but `ibc-app-transfer` does not have these features.
     It has an optional dependency with that name, but that dependency uses the "dep:" syntax in the features table, so it does not have an implicit feature with that name.